test: rename dummGetters
