+++ packages/sanity/src/core/bundles/index.ts
@@ -1,4 +1,4 @@
 export * from './components'
 export * from './hooks'
 export * from './util/const'
-export * from './util/dummyGetters'
+export * from './util/util'
